hid::Button is now an enum class

+++ src/deluge/hid/button.cpp
@@ -7,12 +7,12 @@ namespace button {
 
 struct xy toXY(Button b) {
 #if DELUGE_MODEL == DELUGE_MODEL_40_PAD
-	int x = (unsigned int)b % 10;
-	int y = ((unsigned int)b % 70) / 10;
+	int x = static_cast<unsigned int>(b) % 10;
+	int y = (static_cast<unsigned int>(b) % 70) / 10;
 	y -= displayHeight;
 #else
-	int y = (unsigned int)b / 9;
-	int x = b - y * 9;
+	int y = static_cast<unsigned int>(b) / 9;
+	int x = static_cast<unsigned int>(b) - y * 9;
 	y -= displayHeight * 2;
 #endif
 	return {x, y};
